pyproject.toml
[project]
name = "infolist"
version = "0.1.0"
dependencies = [
    "pyyaml>=6.0"
]
requires-python = ">=3.7,<4.0"